+++ CHANGELOG.md
@@ -14,6 +14,7 @@ $ npm install @ngxs/store@dev
 - Feature(store): Add `updateItems` operator [#2413](https://github.com/ngxs/store/pull/2413)
 - Feature(store): Add `removeItems` operator [#2414](https://github.com/ngxs/store/pull/2414)
 - Feature(store): Expose `ɵgetTypedNgxsStateFactory` for internal third-party usage [#2426](https://github.com/ngxs/store/pull/2426)
+- Feature(store): Error in dev mode when `selectSignal` selector returns new reference with identical value [#2441](https://github.com/ngxs/store/pull/2441)
 - Fix(store): Cleanup observers once subjects complete [#2401](https://github.com/ngxs/store/pull/2401)
 - Fix(store): Skip state mutations when injector is destroyed mid-action [#2406](https://github.com/ngxs/store/pull/2406)
 - Fix(store): Warn when state is mutated after injector destruction [#2407](https://github.com/ngxs/store/pull/2407)